+++ app/src/test/java/de/hohnepeople/keepadb/KeepADBNetworkRobustnessBehaviorTest.java
@@ -19,6 +19,7 @@ import java.lang.reflect.Method;
 import java.net.InetAddress;
 
 import org.junit.After;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
@@ -39,6 +40,9 @@ import org.robolectric.shadows.ShadowWifiInfo;
 @RunWith(RobolectricTestRunner.class)
 @Config(sdk = 34)
 public class KeepADBNetworkRobustnessBehaviorTest {
+
+    @Rule
+    public final KeepADBNetworkResetRule keepADBNetworkResetRule = new KeepADBNetworkResetRule();
 
     private final Context context = RuntimeEnvironment.getApplication();